Changes Success and postman collection uploaded

+++ src/main/java/com/excelr/bank/security/services/impl/BankDataServiceImpl.java
@@ -6,13 +6,13 @@ import com.excelr.bank.models.Bank;
 import com.excelr.bank.payload.response.MessageResponse;
 import com.excelr.bank.repository.BankDataRepository;
 import com.excelr.bank.security.services.interfaces.BankDataService;
-import com.excelr.bank.util.Generator;
 import io.micrometer.common.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 
 
@@ -23,17 +23,29 @@ public class BankDataServiceImpl implements BankDataService {
     private BankDataRepository bankDataRepository;
 
 
-//    public List<Bank> getAllBanks() {
-//        return bankDataRepository.findAll();
-//    }
+    public ResponseEntity<?> getAllBankRecords() {
+        try {
+            List<Bank> bankData = bankDataRepository.findAll();
+            return ResponseEntity.ok(bankData.isEmpty() ? Collections.singletonList("No Records Found") : bankData);
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error Fetching Data");
+        }
+
+    }
 
     @Override
     public ResponseEntity<?> insertRecord(Bank bank) throws BankCreationException {
-        if(isBankValid(bank)) {
-             bankDataRepository.save(bank);
-             return ResponseEntity.status(HttpStatus.OK).body("Record Save Successful");
+        List<Bank> chkBank=bankDataRepository.findByBankName(bank.getBankName());
+        if(chkBank!=null && (chkBank.stream()
+                .anyMatch(rtvBank -> rtvBank.getIfscCode().equals(bank.getIfscCode()))) && (chkBank.stream().anyMatch(rtvBank -> rtvBank.getAddress().equals(bank.getAddress())))) {
+            throw new BankCreationException(bank.getBankName()+ " is already registered with IFSC "+bank.getIfscCode());
         }else{
-            throw new BankCreationException("Bank details are invalid");
+            if (isBankValid(bank)) {
+                bankDataRepository.save(bank);
+                return ResponseEntity.status(HttpStatus.OK).body("Congratulations!!! You are Successfully Registered with ExcelR Banking");
+            } else {
+                throw new BankCreationException("Bank details are invalid");
+            }
         }
     }
     @Override
@@ -56,13 +68,29 @@ public class BankDataServiceImpl implements BankDataService {
     }
 
     @Override
-    public void deleteBank(Long id) {
-        bankDataRepository.deleteById(id);
+    public ResponseEntity<?> deleteBankById(Long id) {
+        Bank bank = bankDataRepository.findById(id).orElseThrow(()-> new RuntimeException("Record Not Available"));
+        if(bank==null){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("No Record Available with Id: "+id);
+        }else{
+            bankDataRepository.deleteById(id);
+            return ResponseEntity.status(HttpStatus.OK).body("Record Delete with Id: "+id);
+        }
+    }
+
+    @Override
+    public ResponseEntity<?> deleteAllBankRecords() {
+        List<Bank> bank = bankDataRepository.findAll();
+        if(bank.isEmpty()){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("No Record Available");
+        }else{
+            bankDataRepository.deleteAll();
+            return ResponseEntity.status(HttpStatus.OK).body("All Data Deleted Successfully");
+        }
     }
 
 
     private boolean isBankValid(Bank bank) {
         return StringUtils.isNotBlank(bank.getBankName());
     }
-
 }
